extract VAI_SYMBOL constant and getReadableFeeVai from Mint
