+++ src/components/v2/MintRepayVai/constants.ts
@@ -1,0 +1,1 @@
+export const VAI_SYMBOL = 'vai';
